Add chat restriction feature

+++ src/commands/delete_all_comments.ts
@@ -2,6 +2,7 @@ import createDebug from 'debug';
 import { client } from '../core';
 import { ChangeStatusEvents } from '../enums';
 import {
+  applyRestrictions,
   autoupdateTaskList,
   changeStatusEvent,
   formatChangeStatusEventMessage,
@@ -16,6 +17,10 @@ const deleteAllCommentsRegex = /^(\/\S+)\s+(\d+)$/;
 
 export const deleteAllTaskComments = async (ctx: Context) => {
   debug('Triggered "delete_all_comments" command');
+
+  if (!(await applyRestrictions(ctx))) {
+    return;
+  }
 
   const message: string = (ctx.message as any).text.trim();
   const match = message.match(deleteAllCommentsRegex);
